+++ src/config/messaging.config.ts
@@ -1,3 +1,4 @@
+import { EWC_CHAIN_ID } from '@energyweb/iam-contracts';
 import { MessagingMethod } from '../modules/messaging/messaging.types';
 import { VOLTA_CHAIN_ID } from '../utils/constants';
 import { ChainId } from './chain.config';
@@ -11,8 +12,13 @@ export interface MessagingConfig {
 const messagingConfig: Record<ChainId, MessagingConfig> = {
   [VOLTA_CHAIN_ID]: {
     messagingMethod: MessagingMethod.Nats,
-    natsServerUrl: 'https://volta-identityevents.energyweb.org/',
+    natsServerUrl: 'https://identityevents-dev.energyweb.org/',
     natsEnvironmentName: 'ewf-volta',
+  },
+  [EWC_CHAIN_ID]: {
+    messagingMethod: MessagingMethod.Nats,
+    natsServerUrl: 'https://identityevents.energyweb.org/',
+    natsEnvironmentName: 'ewf-prod',
   },
 };
 
@@ -20,7 +26,10 @@ const messagingConfig: Record<ChainId, MessagingConfig> = {
  * Used to override existing messaging configuration or add a missing one
  * Configuration must be set before constructing `IAM`
  */
-export const setMessagingConfig = (chainId: ChainId, options: Partial<MessagingConfig>) => {
+export const setMessagingConfig = (
+  chainId: ChainId,
+  options: Partial<MessagingConfig>
+) => {
   messagingConfig[chainId] = { ...messagingConfig[chainId], ...options };
 };