+++ src/utils/detectEnvironment.ts
@@ -4,9 +4,14 @@ export enum ExecutionEnvironment {
 }
 
 export const executionEnvironment = () =>
-  isNode() && !isBrowser() ? ExecutionEnvironment.NODE : ExecutionEnvironment.BROWSER;
+  isNode() && !isBrowser()
+    ? ExecutionEnvironment.NODE
+    : ExecutionEnvironment.BROWSER;
 
 const isNode = () =>
-  typeof process !== 'undefined' && process.versions != null && process.versions.node != null;
+  typeof process !== 'undefined' &&
+  process.versions != null &&
+  process.versions.node != null;
 
-const isBrowser = () => typeof window !== 'undefined' && typeof window.document !== 'undefined';
+const isBrowser = () =>
+  typeof window !== 'undefined' && typeof window.document !== 'undefined';